Move PartialModes.query_subset into SnapshotQuery proper
